Feature/details and fixes

a lot of changes on style and some adjustmen on back end

+++ Frontend/src/app/features/client/clients/clients.component.ts
@@ -1,4 +1,4 @@
-import {Component, inject} from '@angular/core';
+import {Component, inject, ViewChild} from '@angular/core';
 import {MatToolbar} from "@angular/material/toolbar";
 import {MatIconButton, MatMiniFabButton} from "@angular/material/button";
 import {MatIcon} from "@angular/material/icon";
@@ -18,15 +18,27 @@ import {
   MatHeaderCell,
   MatHeaderCellDef,
   MatHeaderRow, MatHeaderRowDef, MatRow, MatRowDef,
-  MatTable
+  MatTable, MatTableDataSource
 } from "@angular/material/table";
-
+import {DialogService} from "../../../core/services/dialog.service";
+import {MatProgressSpinner} from "@angular/material/progress-spinner";
+import {MatSort, MatSortHeader} from "@angular/material/sort";
+import {MatDivider} from "@angular/material/divider";
+import {FaqComponent} from "../../../shared/faq/faq.component";
+import {Extras} from "../../../core/interfaces/extras";
+const FAQ: Extras = {
+  Headline: "FAQ",
+  info: [
+    {
+      title: 'Placeholder question?',
+      message: 'placeholder answer',
+    },
+  ]
+}
 @Component({
   selector: 'app-clients',
   standalone: true,
   imports: [
-    MatToolbar,
-    MatIconButton,
     MatIcon,
     MatFormField,
     MatLabel,
@@ -46,7 +58,12 @@ import {
     MatHeaderRowDef,
     MatRow,
     MatRowDef,
-    MatMiniFabButton
+    MatMiniFabButton,
+    MatProgressSpinner,
+    MatSort,
+    MatSortHeader,
+    MatDivider,
+    FaqComponent
   ],
   templateUrl: './clients.component.html',
   styleUrl: './clients.component.css'
@@ -57,18 +74,31 @@ export class ClientsComponent {
     name: new FormControl("", Validators.required),
     telephone: new FormControl("", Validators.required),
     priceListId: new FormControl("", Validators.required),
+    direction: new FormControl("", Validators.required),
   })
   //service
   private priceListService = inject(PricesService)
   private clientService = inject(ClientService)
+  private dialogService = inject(DialogService)
   //variables
   priceList: PriceList[] = []
-  clientList: Client[] = []
-  columnsToDisplay: string[] = ['Nombre', 'Telefono', 'Lista de Precio', 'Accion']
+  columnsToDisplay: string[] = ['name', 'telephone', 'direction', 'priceListId', 'Accion']
   isUpdating: boolean = false;
   currentEditingId: number = 0;
 
+  isLoading:boolean=false;
+
+  dataSource = new MatTableDataSource<Client>([]);
+
+
+
   //methods
+  @ViewChild(MatSort) sort!: MatSort;
+
+  ngAfterViewInit() {
+    this.dataSource.sort = this.sort;
+  }
+
   ngOnInit() {
     this.loadClientList()
     this.priceListService.getAllPrices().subscribe({
@@ -84,7 +114,8 @@ export class ClientsComponent {
   private loadClientList() {
     this.clientService.getClientList().subscribe({
       next: data => {
-        this.clientList = data
+        // this.clientList = data
+        this.dataSource.data = data
       },
       error: err => {
         console.log(err)
@@ -94,20 +125,17 @@ export class ClientsComponent {
 
 
   save() {
-    console.log(this.clientForm)
     if (this.clientForm.invalid) {
-      console.error(' this form is invalid')
+      alert(' this form is invalid')
     } else {
+      this.isLoading = true;
 
-      if (this.isUpdating){
+      if (this.isUpdating) {
 
         const client: Client = {
           id: this.currentEditingId,
           ...this.clientForm.value
         }
-
-        console.log('enviando datos al backendt', client)
-
         this.clientService.putClient(client).subscribe({
           next: data => {
             console.log('update client')
@@ -117,29 +145,30 @@ export class ClientsComponent {
             this.clientForm.reset()
             this.clientForm.markAsPristine()
             this.clientForm.markAsUntouched()
+            this.isLoading = false
           },
           error: err => {
-            console.log('hubo un error actualizando ',err)
+            console.log('hubo un error actualizando ', err)
+            this.isLoading = false
           }
         })
-      }else {
+      } else {
         const client: Client = {
           ...this.clientForm.value,
         }
-        console.log('enviando datos al backendt', client)
 
         this.clientService.postClient(client).subscribe({
           next: data => {
-            console.log('guardado exitosamente')
-
             this.loadClientList()
 
             this.clientForm.reset()
             this.clientForm.markAsPristine()
             this.clientForm.markAsUntouched()
+            this.isLoading = false
           },
           error: err => {
-            console.log('hubo un error creando la entidad ',err)
+            console.log('hubo un error creando la entidad ', err)
+            this.isLoading = false
           }
         })
       }
@@ -147,19 +176,25 @@ export class ClientsComponent {
   }
 
   delete(id: number) {
-    this.clientService.deleteClient(id).subscribe({
-      next: data => {
-        console.log('Borrado exitosamente')
-        this.loadClientList()
-      },
-      error: err => {
-        console.log(err)
-      }
-    })
+    this.dialogService.confirm('Borrar Cliente', 'esta seguro que desea borrar el cliente permanentemente?')
+      .subscribe(ok => {
+        if (ok) {
+          this.clientService.deleteClient(id).subscribe({
+            next: data => {
+              this.dialogService.alert('Exito', 'Cliente borrado exitosamente').subscribe()
+              this.loadClientList()
+            },
+            error: err => {
+              console.log(err)
+            }
+          })
+        }
+      })
+
   }
 
   edit(id: number) {
-    const clientFilter = this.clientList.find(c => c.id === id);
+    const clientFilter = this.dataSource.data.find(c => c.id === id);
     if (clientFilter) {
       this.isUpdating = true
 
@@ -167,8 +202,9 @@ export class ClientsComponent {
         name: clientFilter.name,
         telephone: clientFilter.telephone,
         priceListId: clientFilter.priceListId,
+        direction: clientFilter.direction,
       })
-      this.currentEditingId= id
+      this.currentEditingId = id
     }
 
   }
@@ -179,12 +215,14 @@ export class ClientsComponent {
   }
 
 
-  getPriceListName(priceListId: number):string {
-    const priceList = this.priceList.find(pl=>pl.id===priceListId)
+  getPriceListName(priceListId: number): string {
+    const priceList = this.priceList.find(pl => pl.id === priceListId)
     if (priceList) {
       return priceList.name
     } else {
       return 'error'
     }
   }
+
+  protected readonly FAQ = FAQ;
 }
